+++ packages/flow-core/src/components/f-accordion/f-accordion.scss
@@ -13,15 +13,8 @@
 			transform: rotate(180deg);
 			transition: transform var(--transition-time-default) linear;
 		}
-
-		&.doc-preview-accordion {
-			padding: 12px 4px 12px 12px !important;
-		}
 	}
 	.f-accordion {
-		&.doc-preview-accordion {
-			padding: 0px 0px 0px 12px !important;
-		}
 		&:not([data-accordion-open]) {
 			max-height: 0 !important;
 			transition: max-height var(--transition-time-default) ease-out;
