+++ packages/flow-core/src/components/f-accordion/f-accordion.ts
@@ -6,7 +6,6 @@ import { FRoot } from "../../mixins/components/f-root/f-root";
 import { FDiv, FDivPaddingProp } from "../f-div/f-div";
 import { flowElement } from "./../../utils";
 import { injectCss } from "@cldcvr/flow-core-config";
-import { classMap } from "lit-html/directives/class-map.js";
 
 injectCss("f-accordion", globalStyle);
 
@@ -147,25 +146,6 @@ export class FAccordion extends FRoot {
 	}
 
 	render() {
-		// classes to apply on inner element
-		const classes: Record<string, boolean> = {
-			"f-accordion": true
-		};
-
-		// classes to apply on inner element
-		const classesHeader: Record<string, boolean> = {
-			"f-accordion-header": true
-		};
-		// merging host classes
-		this.classList.forEach(cl => {
-			classes[cl] = true;
-		});
-
-		// merging host classes
-		this.classList.forEach(cl => {
-			classesHeader[cl] = true;
-		});
-
 		// render accordion-icon
 		const accordionIcon = html` <f-div
 			padding="none none none small"
@@ -189,10 +169,11 @@ export class FAccordion extends FRoot {
 				<f-div
 					height="hug-content"
 					@click=${this.toggleAccordion}
-					align="top-left"
+					align="middle-left"
 					.gap=${this.iconPlacement === "left" ? "medium" : "auto"}
 					width="100%"
-					class=${classMap(classesHeader)}
+					class="f-accordion-header"
+					part="accordion-header"
 					.padding=${this.headerPadding}
 					clickable
 					@mouseenter=${() => this.stateChange("enter")}
@@ -205,7 +186,8 @@ export class FAccordion extends FRoot {
 					${this.icon !== "none" && this.iconPlacement === "right" ? accordionIcon : ""}
 				</f-div>
 				<f-div
-					class=${classMap(classes)}
+					class="f-accordion"
+					part="accordion-body"
 					?data-accordion-open=${this.open}
 					direction="column"
 					.padding=${this.bodyPadding}
