+++ packages/flow-core/src/components/f-document-viewer/f-document-viewer.scss
@@ -38,6 +38,19 @@
 		}
 	}
 
+	f-accordion::part(accordion-header) {
+		padding: 12px 4px 12px 12px !important;
+		justify-content: flex-start;
+		align-items: flex-start;
+	}
+	f-accordion[icon="none"]::part(accordion-header) {
+		padding: 12px 4px 12px 46px !important;
+		justify-content: flex-start;
+		align-items: flex-start;
+	}
+	f-accordion::part(accordion-body) {
+		padding: 0px 0px 0px 12px !important;
+	}
 	#jumplinks-wrapper {
 		overflow-x: hidden !important;
 	}
